move createCaCerts() to multicluster.go
