+++ tests/e2e/framework/multicluster.go
@@ -25,6 +25,7 @@ import (
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"istio.io/istio/pkg/log"
+	"istio.io/istio/tests/util"
 )
 
 func getKubeConfigFromFile(dirname string) (string, error) {
@@ -114,4 +115,23 @@ func (k *KubeInfo) generateRemoteIstio(src, dst string) error {
 		log.Errorf("cannot write remote into generated yaml file %s", dst)
 	}
 	return nil
+}
+
+func (k *KubeInfo) createCacerts(remoteCluster bool) (err error) {
+	kc := k.KubeConfig
+	cluster := "primary"
+	if remoteCluster {
+		kc = k.RemoteKubeConfig
+		cluster = "remote"
+	}
+	caCertFile := filepath.Join(k.ReleaseDir, caCertFileName)
+	caKeyFile := filepath.Join(k.ReleaseDir, caKeyFileName)
+	rootCertFile := filepath.Join(k.ReleaseDir, rootCertFileName)
+	certChainFile := filepath.Join(k.ReleaseDir, certChainFileName)
+	if _, err = util.Shell("kubectl create secret generic cacerts --kubeconfig=%s -n %s "+
+		"--from-file=%s --from-file=%s --from-file=%s --from-file=%s",
+		kc, k.Namespace, caCertFile, caKeyFile, rootCertFile, certChainFile); err == nil {
+		log.Infof("Created Cacerts with namespace %s in %s cluster", k.Namespace, cluster)
+	}
+	return err
 }
